+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add form helpers for security key registration and 2FA authentication.
+
 ## [v0.2.0](https://github.com/cedarcode/devise-webauthn/compare/v0.1.2...v0.2.0/) - 2025-12-03
 
 - Add new `webauthn_two_factor_authenticatable` module for enabling 2FA using WebAuthn credentials.